+++ crates/astria-conductor/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2024-10-25
+
+### Changed
+
+- Bump penumbra dependencies [#1740](https://github.com/astriaorg/astria/pull/1740).
+
 ## [1.0.0-rc.2] - 2024-10-23
 
 ### Changed
@@ -346,7 +352,8 @@ address [#721](https://github.com/astriaorg/astria/pull/721).
 
 - Initial release.
 
-[unreleased]: https://github.com/astriaorg/astria/compare/conductor-v1.0.0-rc.2...HEAD
+[unreleased]: https://github.com/astriaorg/astria/compare/conductor-v1.0.0...HEAD
+[1.0.0]: https://github.com/astriaorg/astria/compare/conductor-v1.0.0-rc.2...conductor-v1.0.0
 [1.0.0-rc.2]: https://github.com/astriaorg/astria/compare/conductor-v1.0.0-rc.1...conductor-v1.0.0-rc.2
 [1.0.0-rc.1]: https://github.com/astriaorg/astria/compare/conductor-v0.20.1...conductor-v1.0.0-rc.1
 [0.20.1]: https://github.com/astriaorg/astria/compare/conductor-v0.20.0...conductor-v0.20.1