+++ crates/astria-core/CHANGELOG.md
@@ -12,3 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial release.
+- Added method `TracePrefixed::leading_channel` to read the left-most channel of
+  a trace prefixed ICS20 asset [#1768](https://github.com/astriaorg/astria/pull/1768)
+
+### Removed
+
+- Removed method `TracePrefixed::last_channel` [#1768](https://github.com/astriaorg/astria/pull/1768)